Switch from pip compile to uv.lock (#116)
Start using a lock file to manage dependencies and ensure
reproducible builds and stable CI jobs.

The uv.lock file is generated via `uv lock`, and it parses the
pyproject.toml file to generate the lock file.

Running uv lock (--check) with an existing `uv.lock` file, checks
that the existing lock file is constistent with `pyproject.toml`
and it does not attempt to update any dependencies.

The `uv lock` command invokes `setup.py` to retrieve metadata.
When this is done on a system without `sendnn` installed, it would
fail. The build script has been udapted so that if `sendnn` is not
found and the current setuptools command is a metadata retrieval
one, the code generation step is skipped and the command does not
fail.

Updated the docs in the requirements/README.md to clarify how to
update requirements and how the requirements files are generated.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>